added new logger implementation

+++ src/presentation/user/user.service.ts
@@ -1,4 +1,4 @@
-import { UpdateUserDto, UserRepository } from "../../domain";
+import { CustomError, UpdateUserDto, UserRepository } from "../../domain";
 
 export class UserService {
   constructor(private userRepository: UserRepository) {}
@@ -10,10 +10,15 @@ export class UserService {
     return await this.userRepository.getUserById(id);
   };
   updateUser = async (updateUserDto: UpdateUserDto) => {
+    const user = await this.getUserById(updateUserDto.id);
+    if (!user)
+      throw CustomError.badRequest(`User does not exists: ${updateUserDto.id}`);
     return await this.userRepository.updateUser(updateUserDto);
   };
 
   deleteUser = async (id: number) => {
+    const user = await this.getUserById(id);
+    if (!user) throw CustomError.badRequest(`User does not exists: ${id}`);
     return await this.userRepository.deleteUser(id);
   };
 }
